requirements.txt
cryptography>=42.0.0
textual>=0.50.0
zeroconf>=0.131.0
netifaces>=0.11.0
python-pkcs11>=0.7.0
aiosqlite>=0.20.0